requirements.txt
moderngl
twisted
pygame
numpy
Pillow
